test(retrieve): add an e2e test for retrieve tool and refactor knowledge_base class and memory tool test
